+++ NEWS.md
@@ -1,2 +1,5 @@
 ## 1.0.0 Unreleased
+ * Added endpoints for loan management at the instance level (currently not
+   supported by FOLIO) for future integration
+ * RAML cleanup
  * Initial work